\name{RecurPart}
\alias{RecurPart}
\alias{RecurPart}
\docType{package}
\title{
\packageTitle{RecurPart}
}
\description{
\packageDescription{RecurPart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RecurPart}
\packageIndices{RecurPart}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RecurPart}

Maintainer: \packageMaintainer{RecurPart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
